feat(#87): extract Cron_Manager from class-sybgo
Move all WP-Cron concerns (schedule registration, hook wiring, callbacks,
deactivation cleanup) into a dedicated Cron_Manager class. Sybgo::init()
delegates to Cron_Manager::init(); Sybgo::deactivate() delegates to the
static Cron_Manager::deactivate(). Uninstaller updated to use
Cron_Manager::get_hooks() instead of the removed Sybgo::get_cron_hooks().
