+++ wp-plugin/class-sybgo.php
@@ -140,7 +140,12 @@ class Sybgo {
 		}
 
 		// Initialize cron schedules.
-		$this->init_cron_schedules();
+		$cron_manager = new Cron_Manager(
+			$this->factory->create_report_manager(),
+			$this->factory->create_email_manager(),
+			$this->factory->create_database_manager()
+		);
+		$cron_manager->init();
 	}
 
 	/**
@@ -334,145 +339,6 @@ class Sybgo {
 	}
 
 	/**
-	 * Get all cron hook names registered by the plugin.
-	 *
-	 * Single source of truth for cron hook names, used both when scheduling
-	 * (init_cron_schedules, deactivate) and when cleaning up (Uninstaller).
-	 *
-	 * @return array<string> List of WP-Cron hook names.
-	 * @since 1.0.0
-	 */
-	public static function get_cron_hooks(): array {
-		return array(
-			'sybgo_freeze_weekly_report',
-			'sybgo_send_report_emails',
-			'sybgo_cleanup_old_events',
-			'sybgo_retry_failed_emails',
-		);
-	}
-
-	/**
-	 * Initialize cron schedules.
-	 *
-	 * @return void
-	 */
-	private function init_cron_schedules(): void {
-		$hooks = self::get_cron_hooks();
-
-		// Register custom cron intervals.
-		add_filter( 'cron_schedules', array( $this, 'add_cron_intervals' ) );
-
-		// Schedule weekly freeze (Sunday 23:55).
-		if ( ! wp_next_scheduled( $hooks[0] ) ) {
-			$next_sunday = strtotime( 'next Sunday 23:55' );
-			wp_schedule_event( $next_sunday, 'weekly', $hooks[0] );
-		}
-
-		// Schedule weekly email (Monday 00:05).
-		if ( ! wp_next_scheduled( $hooks[1] ) ) {
-			$next_monday = strtotime( 'next Monday 00:05' );
-			wp_schedule_event( $next_monday, 'weekly', $hooks[1] );
-		}
-
-		// Schedule daily cleanup (3am).
-		if ( ! wp_next_scheduled( $hooks[2] ) ) {
-			$next_3am = strtotime( 'tomorrow 3:00' );
-			wp_schedule_event( $next_3am, 'daily', $hooks[2] );
-		}
-
-		// Schedule daily retry failed emails (9am).
-		if ( ! wp_next_scheduled( $hooks[3] ) ) {
-			$next_9am = strtotime( 'tomorrow 9:00' );
-			wp_schedule_event( $next_9am, 'daily', $hooks[3] );
-		}
-
-		// Register cron callbacks.
-		add_action( $hooks[0], array( $this, 'freeze_weekly_report_callback' ) );
-		add_action( $hooks[1], array( $this, 'send_report_emails_callback' ) );
-		add_action( $hooks[2], array( $this, 'cleanup_old_events_callback' ) );
-		add_action( $hooks[3], array( $this, 'retry_failed_emails_callback' ) );
-	}
-
-	/**
-	 * Add custom cron intervals.
-	 *
-	 * @param array<string, array<string, mixed>> $schedules Existing schedules.
-	 * @return array<string, array<string, mixed>> Modified schedules.
-	 */
-	public function add_cron_intervals( array $schedules ): array {
-		if ( ! isset( $schedules['weekly'] ) ) {
-			$schedules['weekly'] = array(
-				'interval' => 604800, // 7 days in seconds.
-				'display'  => esc_html__( 'Once Weekly', 'sybgo' ),
-			);
-		}
-		return $schedules;
-	}
-
-	/**
-	 * Cron callback: Freeze weekly report.
-	 *
-	 * @return void
-	 */
-	public function freeze_weekly_report_callback(): void {
-		$report_manager = $this->factory->create_report_manager();
-		$frozen_id      = $report_manager->freeze_current_report();
-
-		if ( $frozen_id ) {
-			Logger::info( sprintf( 'Weekly report #%d frozen successfully', $frozen_id ) );
-		} else {
-			Logger::info( 'No active report to freeze' );
-		}
-	}
-
-	/**
-	 * Cron callback: Send report emails.
-	 *
-	 * @return void
-	 */
-	public function send_report_emails_callback(): void {
-		$report_repo   = $this->factory->create_report_repository();
-		$email_manager = $this->factory->create_email_manager();
-
-		// Get last frozen report.
-		$last_frozen = $report_repo->get_last_frozen();
-
-		if ( ! $last_frozen ) {
-			return;
-		}
-
-		// Cast report id to int — $wpdb returns column values as strings, but
-		// Email_Manager::send_report_email() is strictly typed against int.
-		$report_id = (int) $last_frozen['id'];
-
-		// Send email.
-		$sent = $email_manager->send_report_email( $report_id );
-
-		// Log result.
-		if ( $sent ) {
-			Logger::info( sprintf( 'Successfully sent weekly digest for report #%d', $report_id ) );
-		} else {
-			Logger::error( sprintf( 'Failed to send weekly digest for report #%d', $report_id ) );
-		}
-	}
-
-	/**
-	 * Cron callback: Cleanup old events.
-	 *
-	 * @return void
-	 */
-	public function cleanup_old_events_callback(): void {
-		$db_manager = $this->factory->create_database_manager();
-		$days       = Admin\Settings_Page::get_retention_days();
-		$deleted    = $db_manager->cleanup_old_events( $days );
-
-		// Log cleanup action.
-		if ( $deleted > 0 ) {
-			Logger::info( sprintf( 'Cleaned up %d rows (retention: %d days)', $deleted, $days ) );
-		}
-	}
-
-	/**
 	 * Handle manual cleanup form submission.
 	 *
 	 * Verifies nonce and capability, runs cleanup with the configured retention period,
@@ -512,23 +378,6 @@ class Sybgo {
 			)
 		);
 		exit;
-	}
-
-	/**
-	 * Cron callback: Retry failed emails.
-	 *
-	 * @return void
-	 */
-	public function retry_failed_emails_callback(): void {
-		$email_manager = $this->factory->create_email_manager();
-
-		// Retry failed emails.
-		$retried = $email_manager->retry_failed_emails();
-
-		// Log result.
-		if ( $retried > 0 ) {
-			Logger::info( sprintf( 'Retried %d failed emails', $retried ) );
-		}
 	}
 
 	/**
@@ -610,10 +459,7 @@ class Sybgo {
 	 * @return void
 	 */
 	public function deactivate(): void {
-		// Clear scheduled events.
-		foreach ( self::get_cron_hooks() as $hook ) {
-			wp_clear_scheduled_hook( $hook );
-		}
+		Cron_Manager::deactivate();
 
 		// Flush rewrite rules.
 		flush_rewrite_rules();
